qa: resolve Psalm errors
- Pushes test functions to a separate file and adds autoloading for it
  - Also excludes this file from Psalm analysis
- Removes unnecessary array_values call
