+++ test/DefinitionTest.php
@@ -81,7 +81,7 @@ class DefinitionTest extends TestCase
         $this->assertNotEquals($methods, $this->definition->getMethods());
         $this->definition->setMethods($methods);
         $test = $this->definition->getMethods();
-        $this->assertEquals(array_values($methods), array_values($test));
+        $this->assertEquals($methods, array_values($test));
     }
 
     public function testHasMethodShouldReturnFalseWhenMethodNotRegisteredWithDefinition(): void
